feat: deprecate old auth
