+++ cmd/mqtt-bridge/mqtt-bridge.go
@@ -141,7 +141,7 @@ func main() {
 	}
 
 	SIGNING_KEY := []byte(viper.GetString("SIGNING_KEY"))
-	auth.SetContext(log, nil, nil, SIGNING_KEY)
+	auth := auth.NewAuthInterceptor(log, nil, nil, SIGNING_KEY)
 	token, err := auth.MakeToken(schema.ROOT_ACCOUNT_KEY)
 	if err != nil {
 		log.Fatal("Error making token", zap.Error(err))
